💄 [style] #472 : eslint에 맞춰서 리펙토링
