+++ server/.eslintrc.json
@@ -25,7 +25,7 @@
     //    "import/extensions": "off",
   },
 
-  "ignorePatterns": [".eslintrc.json"]
+  "ignorePatterns": [".eslintrc.json", "test"]
 }
 
 /* Nest.js 기본 eslint 설정 */
