+++ server/src/domain/alarms/alarms.controller.ts
@@ -1,21 +1,15 @@
-import { JwtAuthGuard } from "@guard/jwt.guard";
 import { Controller, Get, Req, Query, Sse } from "@nestjs/common";
 import { ApiOperation, ApiQuery, ApiTags } from "@nestjs/swagger";
 import { Exception } from "@exception/exceptions";
 import { isValidUserId } from "@validation/validation";
-import { UsersService } from "@user/users.service";
 import { interval, map } from "rxjs";
-import { GetUserId, NoAuth } from "../../decorator/validate.decorator";
+import { RequestWithUser } from "@type/request";
 import { AlarmsService } from "./alarms.service";
-import { RequestWithUser } from "../../types/request.d";
 
 @Controller("api/alarms")
 @ApiTags("ALARM API")
 export class AlarmsController {
-  constructor(
-    private readonly alarmService: AlarmsService,
-    private readonly usersService: UsersService,
-  ) {}
+  constructor(private readonly alarmService: AlarmsService) {}
 
   @Get("static/unread")
   @ApiOperation({
@@ -43,7 +37,6 @@ export class AlarmsController {
     return interval(1000).pipe(
       map(() => {
         const userId = Number(req.user);
-        console.log(userId, global.alarmBar);
         let fetchAlarmSign = false;
         if (global.alarmBar.has(userId)) {
           fetchAlarmSign = true;
