+++ server/src/domain/event/event.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable, Scope } from "@nestjs/common";
+import { Injectable } from "@nestjs/common";
 import EventEmitter from "events";
 import { fromEvent } from "rxjs";
 
@@ -20,6 +20,7 @@ export class EventService {
       if (isUserConnect) {
         this.emitter.emit(`${userId}ch`, { data: "true" });
       }
+      return null;
     });
   }
 }
