+++ server/src/domain/exercises/dto/exercise.dto.ts
@@ -1,16 +1,7 @@
-import {
-  ArrayNotEmpty,
-  IsArray,
-  IsNotEmpty,
-  IsNumber,
-  IsString,
-  Matches,
-  Max,
-  Min,
-  ValidateNested,
-} from "class-validator";
+import { ArrayNotEmpty, IsArray, IsNotEmpty, IsNumber, Min, ValidateNested } from "class-validator";
 import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
+import { SingleExercise } from "@exercise/dto/single-exercise.dto";
 
 export class ExerciseDataDto {
   @ApiProperty({
@@ -31,56 +22,4 @@ export class ExerciseDataDto {
   @Type(() => SingleExercise)
   @ValidateNested({ each: true })
   exerciseList: SingleExercise[];
-}
-
-export class SingleExercise {
-  @ApiProperty({
-    description: "운동 이름",
-    type: String,
-  })
-  @IsNotEmpty()
-  @IsString()
-  exerciseName: string;
-
-  @ApiProperty({
-    description: "운동 배열",
-    type: () => [SingleSet],
-  })
-  @IsArray()
-  @ArrayNotEmpty()
-  @Type(() => SingleSet)
-  @ValidateNested({ each: true })
-  setList: SingleSet[];
-}
-
-export class SingleSet {
-  @ApiProperty({
-    description: "중량",
-    type: Number,
-  })
-  @IsNotEmpty()
-  @IsNumber()
-  @Min(0)
-  @Max(1000)
-  kg: number;
-
-  @ApiProperty({
-    description: "반복 횟수",
-    type: Number,
-  })
-  @IsNotEmpty()
-  @IsNumber()
-  @Min(1)
-  @Max(100)
-  count: number;
-
-  @ApiProperty({
-    description: "수행 여부",
-    type: Number,
-  })
-  @IsNotEmpty()
-  @IsNumber()
-  @Min(0)
-  @Max(1)
-  check: number;
 }
