+++ server/src/domain/exercises/dto/single-exercise.dto.ts
@@ -1,0 +1,24 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { ArrayNotEmpty, IsArray, IsNotEmpty, IsString, ValidateNested } from "class-validator";
+import { Type } from "class-transformer";
+import { SingleSet } from "@exercise/dto/single-set.dto";
+
+export class SingleExercise {
+  @ApiProperty({
+    description: "운동 이름",
+    type: String,
+  })
+  @IsNotEmpty()
+  @IsString()
+  exerciseName: string;
+
+  @ApiProperty({
+    description: "운동 배열",
+    type: () => [SingleSet],
+  })
+  @IsArray()
+  @ArrayNotEmpty()
+  @Type(() => SingleSet)
+  @ValidateNested({ each: true })
+  setList: SingleSet[];
+}
