+++ server/src/domain/exercises/dto/single-set.dto.ts
@@ -1,0 +1,34 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { IsNotEmpty, IsNumber, Max, Min } from "class-validator";
+
+export class SingleSet {
+  @ApiProperty({
+    description: "중량",
+    type: Number,
+  })
+  @IsNotEmpty()
+  @IsNumber()
+  @Min(0)
+  @Max(1000)
+  kg: number;
+
+  @ApiProperty({
+    description: "반복 횟수",
+    type: Number,
+  })
+  @IsNotEmpty()
+  @IsNumber()
+  @Min(1)
+  @Max(100)
+  count: number;
+
+  @ApiProperty({
+    description: "수행 여부",
+    type: Number,
+  })
+  @IsNotEmpty()
+  @IsNumber()
+  @Min(0)
+  @Max(1)
+  check: number;
+}
