+++ server/src/domain/mock/mock.controller.ts
@@ -25,7 +25,7 @@ export class MockController {
       }),
     );
 
-    // eslint-disable-next-line no-restricted-syntax
+    // eslint-disable-next-line no-restricted-syntax,@typescript-eslint/no-unused-vars
     for (const id of userNums) {
       // eslint-disable-next-line no-await-in-loop
       await this.mockService.mockUsers();
