+++ server/src/domain/routines/dto/single-routine.dto.ts
@@ -1,16 +1,7 @@
-import {
-  ArrayNotEmpty,
-  IsArray,
-  IsNotEmpty,
-  IsNumber,
-  IsOptional,
-  IsString,
-  Max,
-  Min,
-  ValidateNested,
-} from "class-validator";
+import { ArrayNotEmpty, IsArray, IsNumber, IsString, Min, ValidateNested } from "class-validator";
 import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
+import { SingleExercise } from "@routine/dto/single-exercise.dto";
 
 export class RoutineDto {
   @ApiProperty({
@@ -37,54 +28,4 @@ export class RoutineDto {
   @Type(() => SingleExercise)
   @ValidateNested({ each: true })
   exerciseList: SingleExercise[];
-}
-
-export class SingleExercise {
-  @ApiProperty({
-    description: "루틴 고유 ID",
-    type: Number,
-  })
-  @IsNumber()
-  @IsOptional()
-  routineId: number;
-
-  @ApiProperty({
-    description: "운동 이름",
-    type: String,
-  })
-  @IsNotEmpty()
-  @IsString()
-  exerciseName: string;
-
-  @ApiProperty({
-    description: "운동 배열",
-    type: () => [SingleSet],
-  })
-  @IsArray()
-  @ArrayNotEmpty()
-  @Type(() => SingleSet)
-  @ValidateNested({ each: true })
-  setList: SingleSet[];
-}
-
-export class SingleSet {
-  @ApiProperty({
-    description: "중량",
-    type: Number,
-  })
-  @IsNotEmpty()
-  @IsNumber()
-  @Min(0)
-  @Max(1000)
-  kg: number;
-
-  @ApiProperty({
-    description: "반복 횟수",
-    type: Number,
-  })
-  @IsNotEmpty()
-  @IsNumber()
-  @Min(0)
-  @Max(100)
-  count: number;
 }
