+++ server/src/domain/sbd_records/converter/sbd_records.converter.ts
@@ -16,6 +16,7 @@ export const recordConverter = {
           userWeight: record.userWeight,
         },
       });
+      return recordObject;
     });
     return recordObject;
   },
