+++ server/src/domain/sbd_records/dto/single_sbd_data.dto.ts
@@ -1,4 +1,4 @@
-import { IsNumber, IsString, Length, Matches, Min } from "class-validator";
+import { IsNumber, Min } from "class-validator";
 import { ApiProperty } from "@nestjs/swagger";
 
 export class SingleSBDDataDto {
