+++ server/src/guard/jwt.strategy.ts
@@ -1,7 +1,7 @@
 import { Strategy } from "passport-jwt";
 import { PassportStrategy } from "@nestjs/passport";
 import { Injectable } from "@nestjs/common";
-import { Request, Response } from "express";
+import { Request } from "express";
 import { ACCESS_TOKEN_SECRETKEY } from "@utils/env";
 import { JwtPayload } from "@type/jwt";
 import { Exception } from "@exception/exceptions";
